BattleDiamondSpawnerQSystem merged into BattleDiamondQSystem
