+++ Assets/QuantumUser/Resources/Configs/QuantumBattleSystemsConfig.asset
@@ -103,12 +103,6 @@ MonoBehaviour:
     StartDisabled: 0
     Children: []
   - SystemType:
-      AssemblyQualifiedName: Battle.QSimulation.Diamond.BattleDiamondSpawnerQSystem,
-        Quantum.Simulation
-    SystemName: 
-    StartDisabled: 0
-    Children: []
-  - SystemType:
       AssemblyQualifiedName: Battle.QSimulation.Diamond.BattleDiamondQSystem, Quantum.Simulation
     SystemName: 
     StartDisabled: 0
