--- Assets/QuantumUser/Simulation/Battle/Scripts/Diamonds/BattleDiamondSpawnerQSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a1bcded0f8b84b1a7b5c4d2b0f51914e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
